python/requirements.txt
pyopengl
glfw
